Add new class for SortedOrderedDict
The new class now handles ordering dict entries on insertion as well as
having a maximum length (like a deque would). Practically speaking what
I'd really like is some way to subclass OrderedDict (or even regular
dict) and then mix in properties of deques and defaultdicts but I don't
know how (if possible) to do that.