Add count aggregation methods, lots of smaller cleanup
Add a new method each to CountersCollection and CounterSeries to allow
aggregation over time and metric names, which we'll be using for the
summaries and alerting.

Some general things also happening here: adding more tests, fixing
docstrings, fixing types, and (most importantly after the new methods)
changing behavior so the parser is the only class responsible for
taking in a string and turning it into a datetime.